fix: bound example key names in validation error message
Wrap unknownKeys.join() with short() to prevent unbounded key names
from appearing in error messages when a caller provides objects with
very long property names. Matches the existing pattern used for all
other caller-supplied text in this validator.

Add test asserting a ceiling on error message length for hostile inputs.

+++ src/server.ts
@@ -274,7 +274,7 @@ export async function handleToolCall(
       const unknownKeys = Object.keys(ex as object).filter((k) => k !== "input" && k !== "output");
       if (unknownKeys.length > 0) {
         throw new Error(
-          `\`examples[${i}]\` has unrecognized key(s): ${unknownKeys.join(", ")}. ` +
+          `\`examples[${i}]\` has unrecognized key(s): ${short(unknownKeys.join(", "))}. ` +
             `Only \`input\` and \`output\` are valid; got: ${short(ex)}`,
         );
       }
